Revert "ignore changes in submodules when committing everything"
This reverts commit d78ac04ee609b0b63bb688be2a837a6ca7670392.

The benefits of `ignore=all` are that submodules are omitted from
`git status` and `git citool`. The drawbacks are - submodules
are not omitted from `git commit -a` *and* they are omitted
from `git diff` and `git show` output.

As a result one can unintentionally commit changes to submodules and not
see it in the history. Thus drawbacks outweigh benefits here.